Add quick-nav component for revealing structure of large pages
This commit adds a small drop-down menu at the top of each page. When clicked,
it reveals an in-page hierarchical TOC which you can use to jump directly to any
of the H2, H3, or H4 headers in the primary content. The TOC can be dismissed by
clicking outside it or by re-clicking the trigger.

This re-uses the SCSS from the top-nav dropdown, and uses a little bit of JQuery
to do the rest of the lifting. I ruled out using Ruby to get the list of headers
during the static build for a couple of reasons:

- It would be significantly harder to accomplish.
- It would be a dead-end investment, since we're moving away from the Middleman
  platform.
- It risks a nasty effect on site build times and/or HTML munging code.
  (Previous experiments at using Nogokiri to safely extract what I want from the
  content HTML have not been good, but doing it with regexps can be even worse.)

The extra event handlers to make click-based activation/deactivation work
wouldn't be necessary if I'd just used hover-based activation like in the top
navs, but that felt awkward once the TOC reached about the length of a screen.